moved the data files into a same folder
