+++ src/components/Card.js
@@ -1,23 +1,36 @@
-export default function Card(props) {
-  let badgeText
-      if (props.openSpots === 0) {
-        badgeText = "SOLD OUT"
-      } else if (props.platform === "Online") {
-        badgeText = "ONLINE"
-      }
+export default function Card({
+  openSpots,
+  platform,
+  img,
+  star,
+  reviews,
+  location,
+  title,
+  price,
+  rating,
+}) {
+  let badgeText;
+  if (openSpots === 0) {
+    badgeText = "SOLD OUT";
+  } else if (platform === "Online") {
+    badgeText = "ONLINE";
+  }
   return (
     <section className="cards">
       {badgeText && <p className="availability">{badgeText}</p>}
-      <img className="cards--pic" src={props.img} alt="Olympic Swimmer" />
+      <img className="cards--pic" src={img} alt="Olympic Swimmer" />
       <div>
-        <img className="cards--star" src={props.star} alt="" />
+        <img className="cards--star" src={star} alt="" />
         <p className="cards--text">
-          {props.rating} <span className="text--location">({props.reviews}) • {props.location}</span>
+          {rating}{" "}
+          <span className="text--location">
+            ({reviews}) • {location}
+          </span>
         </p>
       </div>
-      <p className="cards--text">{props.title}</p>
+      <p className="cards--text">{title}</p>
       <p className="cards--price">
-        <strong>From ${props.price}</strong> / person
+        <strong>From ${price}</strong> / person
       </p>
     </section>
   );
